updated to CMSimple_XH 1.6.4

+++ www/plugins/filebrowser/classes/Filebrowser_Controller.php
@@ -11,7 +11,7 @@
  * @author    The CMSimple_XH developers <devs@cmsimple-xh.org>
  * @copyright 2009-2014 The CMSimple_XH developers <http://cmsimple-xh.org/?The_Team>
  * @license   http://www.gnu.org/licenses/gpl-3.0.en.html GNU GPLv3
- * @version   SVN: $Id: Filebrowser_Controller.php 1198 2014-01-30 14:10:39Z cmb69 $
+ * @version   SVN: $Id: Filebrowser_Controller.php 1379 2014-09-21 19:47:28Z cmb69 $
  * @link      http://cmsimple-xh.org/
  */
 
@@ -49,6 +49,13 @@ class Filebrowser_Controller
      * @var string $currentDirectory
      */
     var $currentDirectory;
+
+    /**
+     *  The current type of allowed file extensions.
+     *
+     *  @var string
+     */
+    var $currentType;
 
     /**
      * @var string $linkType
@@ -117,6 +124,29 @@ class Filebrowser_Controller
                 if ((bool) $extension) {
                     $this->allowedExtensions[$type][] = strtolower($extension);
                 }
+            }
+        }
+    }
+
+    /**
+     * Determines the current type of the allowed file extensions.
+     *
+     * @return void
+     */
+    function determineCurrentType()
+    {
+        $this->currentType = $this->linkType;
+        if (count(array_unique($this->baseDirectories)) != 4) {
+            // If any of the directories are identical, we can't reliably detect
+            // the current type, so we bail out.
+            return;
+        }
+        $types = array('images', 'downloads', 'media', 'userfiles');
+        foreach ($types as $type) {
+            $pos = strpos($this->currentDirectory, $this->baseDirectories[$type]);
+            if ($pos === 0) {
+                $this->currentType = $type;
+                break;
             }
         }
     }
@@ -266,8 +296,8 @@ class Filebrowser_Controller
         if ($extension == $file) {
             return false;
         }
-        if (!in_array($extension, $this->allowedExtensions[$this->linkType])
-            && !in_array('*', $this->allowedExtensions[$this->linkType])
+        if (!in_array($extension, $this->allowedExtensions[$this->currentType])
+            && !in_array('*', $this->allowedExtensions[$this->currentType])
         ) {
             return false;
         }
@@ -398,19 +428,24 @@ class Filebrowser_Controller
     {
         $file = $_FILES['fbupload'];
 
-        if ($file['error'] != 0) {
-            switch ($file['error']) {
-            case UPLOAD_ERR_INI_SIZE:
-                $this->view->error('error_not_uploaded', $file['name']);
-                $this->view->info(
-                    'error_file_too_big_php',
-                    array(ini_get('upload_max_filesize'), 'upload_max_filesize')
-                );
-                return;
-            default:
-                $this->view->error('error_not_uploaded', $file['name']);
-                return;
-            }
+        switch ($file['error']) {
+        case UPLOAD_ERR_OK:
+            break;
+        case UPLOAD_ERR_INI_SIZE:
+            $this->view->error(
+                'error_file_too_big_php',
+                array(ini_get('upload_max_filesize'), 'upload_max_filesize')
+            );
+            $this->view->info('error_not_uploaded', $file['name']);
+            return;
+        case UPLOAD_ERR_NO_TMP_DIR:
+            $this->view->error('error_missing_temp_folder');
+            $this->view->info('error_not_uploaded', $file['name']);
+            return;
+        default:
+            $this->view->error('error_unknown', (string) $file['error']);
+            $this->view->info('error_not_uploaded', $file['name']);
+            return;
         }
 
         // alternatively the following might be used:
@@ -482,7 +517,7 @@ class Filebrowser_Controller
             return;
         }
         if (!mkdir($folder)) {
-            $this->view->error('error_unknown');
+            $this->view->error('error_cant_create_folder');
         }
         $this->view->success('success_folder_created', basename($folder));
     }
@@ -496,11 +531,38 @@ class Filebrowser_Controller
     {
         $folder = $this->browseBase . $this->currentDirectory
             . basename($_POST['folder']);
-        if (!rmdir($folder)) {
-            $this->view->error('error_not_deleted', basename($folder));
+        if (!$this->_isEmpty($folder)) {
+            $this->view->error('error_folder_not_empty');
+            $this->view->info('error_not_deleted', basename($folder));
             return;
+        } else {
+            if (!rmdir($folder)) {
+                $this->view->error('error_not_deleted', basename($folder));
+                return;
+            }
         }
         $this->view->success('success_deleted', basename($folder));
+    }
+
+    /**
+     * Returns whether a folder is empty.
+     *
+     * @param string $folder A folder name.
+     *
+     * @return bool
+     */
+    function _isEmpty($folder)
+    {
+        $isEmpty = true;
+        if ($dir = opendir($folder)) {
+            while (($entry = readdir($dir)) !== false) {
+                if ($entry != '.' && $entry != '..') {
+                    $isEmpty = false;
+                    break;
+                }
+            }
+        }
+        return $isEmpty;
     }
 
     /**
